+++ src/Control/Lens/Grammar/Token.hs
@@ -20,11 +20,18 @@ module Control.Lens.Grammar.Token
 
 import Control.Lens
 import Control.Lens.PartialIso
+import Control.Monad.Loops (iterateUntil)
 import Data.Bifunctor.Joker
 import Data.Char
+import Data.Foldable
 import Data.Profunctor
 import Data.Profunctor.Monoidal
 import Data.Word
+import Control.Monad.State (StateT, state)
+import System.Random (RandomGen, Random, random, randomR)
+import Test.QuickCheck.Arbitrary
+import Test.QuickCheck.Gen (Gen)
+import qualified Test.QuickCheck.Gen as Gen
 import Text.ParserCombinators.ReadP (ReadP)
 import qualified Text.ParserCombinators.ReadP as ReadP
 
@@ -137,3 +144,21 @@ instance Tokenized Char (ReadP Char) where
   notOneOf = ReadP.satisfy . notOneOf
   asIn = ReadP.satisfy . asIn
   notAsIn = ReadP.satisfy . notAsIn
+instance (Categorized token, Arbitrary token) => Tokenized token (Gen token) where
+  anyToken = arbitrary @token
+  token = pure
+  oneOf = Gen.elements . toList
+  notOneOf xs = arbitrary `Gen.suchThat` (`notElem` xs)
+  asIn cat = arbitrary `Gen.suchThat` ((==) cat . categorize)
+  notAsIn cat = arbitrary `Gen.suchThat` ((/=) cat . categorize)
+instance (RandomGen g, Monad m, Categorized token, Random token)
+  => Tokenized token (StateT g m token) where
+  anyToken = state random
+  token = pure
+  oneOf xs = do
+    let ys = toList xs
+    i <- state (randomR (0, length ys - 1))
+    pure (ys !! i)
+  notOneOf xs = iterateUntil (`notElem` xs) anyToken
+  asIn cat = iterateUntil ((== cat) . categorize) anyToken
+  notAsIn cat = iterateUntil ((/= cat) . categorize) anyToken